now correct dtype

+++ code/rnn-pytorch-tree-teacher.py
@@ -158,9 +158,9 @@ loss_func = nn.L1Loss()
 
 out_train = {}
 for appliance in ORDER:
-    out_train[appliance] = Variable(torch.Tensor(eval("train_"+appliance).reshape((train_agg.shape[0], -1, 1))).astype(dtype))
+    out_train[appliance] = Variable(torch.Tensor(eval("train_"+appliance).reshape((train_agg.shape[0], -1, 1))).type(dtype))
 
-inp = Variable(torch.Tensor(train_agg.reshape((train_agg.shape[0], -1, 1))).astype(dtype), requires_grad=True)
+inp = Variable(torch.Tensor(train_agg.reshape((train_agg.shape[0], -1, 1))).type(dtype), requires_grad=True)
 
 for t in range(num_iterations):
